update callback location
