+++ swanlab/data/callback_cloud.py
@@ -26,6 +26,7 @@ from .callback_local import LocalRunCallback, get_run, SwanLabRunState
 from swanlab.cloud import LogSnifferTask, ThreadPool
 from swanlab.db import Experiment
 from swanlab.utils import create_time
+from swanlab.package import get_package_version, get_package_lastest_version
 import sys
 import os
 import io
@@ -96,6 +97,15 @@ class CloudRunCallback(LocalRunCallback):
         if tp != KeyboardInterrupt:
             raise tp(val)
 
+    def _lastest_version_print(self):
+        """
+        cloud模式训练开始时，检测package是否为最新版本
+        """
+        lastest_version = get_package_lastest_version()
+        local_version = get_package_version()
+        if lastest_version is not None and lastest_version != local_version:
+            swanlog.info(f"swanlab version {lastest_version} is available!  Upgrade: `pip install -U swanlab`")
+
     def __str__(self):
         return "SwanLabCloudRunCallback"
 
@@ -104,7 +114,12 @@ class CloudRunCallback(LocalRunCallback):
         if self.login_info is None:
             swanlog.debug("Login info is None, get login info.")
             self.login_info = self.get_login_info()
+
         http = create_http(self.login_info)
+
+        # 检测是否有最新的版本
+        self._lastest_version_print()
+
         return http.mount_project(project, workspace).history_exp_count
 
     def on_run(self):
@@ -133,8 +148,6 @@ class CloudRunCallback(LocalRunCallback):
 
         swanlog.set_write_callback(_write_call_call)
 
-        # 检测是否有最新的版本
-        self._lastest_version_print()
         # 注册系统回调
         self._register_sys_callback()
         # 打印信息
